backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
tinydb==4.8.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
reportlab==4.0.4
Flask-SQLAlchemy==3.0.5
yfinance==0.2.18
plotly==5.17.0
